Move message stuff into own folder
