+++ ui/src/Layout.tsx
@@ -13,7 +13,7 @@ import * as config from './config';
 import Applications from './application/Applications';
 import Clients from './client/Clients';
 import Login from './pages/Login';
-import Messages from './pages/Messages';
+import Messages from './message/Messages';
 import Users from './pages/Users';
 import {observer} from 'mobx-react';
 import {observable} from 'mobx';
